+++ metadata-ingestion/tests/unit/utilities/test_cli_logging.py
